Chest search implemented
Now whenever a player closes a chest in a shop, the plugin will
calculate what items the shop has in stock. Players can then search for
these items with /searchshops <item>.
Placing a sign with the words "supply","supplies", or "stock" next to a
chest causes the search to skip that chest. Signs are very finicky, so
I'm currently trying to fix them.